+++ R/zzz.R
@@ -1,4 +1,8 @@
-utils::globalVariables(c("G","E"))
+.onLoad <- function(libname, pkgname) {
+  utils::globalVariables(c("G","E"))
+  assign(".DollarNames.ABM_G", .DollarNames.ABM_G, envir = globalenv())
+  assign(".DollarNames.ABM_Agent", .DollarNames.ABM_Agent, envir = globalenv())
+}
 
 #' @keywords internal
 .DollarNames.ABM_G <- function(x, pattern = "") {